refactor(flags): Rename new-flags to flags and update import paths
- Rename `src/new-flags.ts` to `src/flags.ts`
- Rename `src/new-flags.spec.ts` to `src/flags.spec.ts`
- Update import paths in all sample files from `../src/new-flags` to `../src/flags`
- Update tsconfig.base.json to exclude samples directory
- Maintain existing functionality and type definitions
- Simplify project structure by removing temporary naming convention
